requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
email-validator>=1.1.3
dnspython>=2.1.0
pytest>=6.2.5
httpx>=0.23.0  
slowapi>=0.1.4
pydantic>=1.8.2
python-dotenv
requests